backend/requirements.txt
fastapi==0.116.1
pandas==2.2.2
joblib==1.4.2
python-dotenv
requests==2.31.0
motor
uvicorn==0.35.0
pydantic==2.11.2
xgboost==3.1.2
scikit-learn==1.6.1